# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SentinelGPT API",
    "description": "SentinelGPT: AI-Powered Vulnerability Analysis Platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/schedules": {
      "post": {
        "tags": [
          "Schedules"
        ],
        "summary": "Create a recurring scan schedule",
        "description": "Schedule authorized rescans of an owned target on an interval.",
        "operationId": "create_schedule_api_v1_schedules_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduleRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Schedules"
        ],
        "summary": "List owned schedules",
        "operationId": "list_schedules_api_v1_schedules_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScheduleResponse"
                  },
                  "title": "Response List Schedules Api V1 Schedules Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/schedules/{schedule_id}": {
      "get": {
        "tags": [
          "Schedules"
        ],
        "summary": "Get schedule detail",
        "operationId": "get_schedule_api_v1_schedules__schedule_id__get",
        "parameters": [
          {
            "name": "schedule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Schedule Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Schedules"
        ],
        "summary": "Update a schedule",
        "operationId": "update_schedule_api_v1_schedules__schedule_id__patch",
        "parameters": [
          {
            "name": "schedule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Schedule Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScheduleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Schedules"
        ],
        "summary": "Delete a schedule",
        "operationId": "delete_schedule_api_v1_schedules__schedule_id__delete",
        "parameters": [
          {
            "name": "schedule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Schedule Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/schedules/{schedule_id}/trigger": {
      "post": {
        "tags": [
          "Schedules"
        ],
        "summary": "Run one schedule tick immediately",
        "description": "Execute one gated tick now (same gates as automatic runs).",
        "operationId": "trigger_schedule_api_v1_schedules__schedule_id__trigger_post",
        "parameters": [
          {
            "name": "schedule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Schedule Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateScheduleRequest": {
        "properties": {
          "targetId": {
            "type": "string",
            "format": "uuid",
            "title": "Targetid"
          },
          "scanProfile": {
            "type": "string",
            "maxLength": 50,
            "title": "Scanprofile",
            "default": "standard"
          },
          "intervalSeconds": {
            "type": "integer",
            "title": "Intervalseconds"
          }
        },
        "type": "object",
        "required": [
          "targetId",
          "intervalSeconds"
        ],
        "title": "CreateScheduleRequest",
        "description": "POST /schedules request body."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScheduleResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "targetId": {
            "type": "string",
            "format": "uuid",
            "title": "Targetid"
          },
          "scanProfile": {
            "type": "string",
            "title": "Scanprofile"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "intervalSeconds": {
            "type": "integer",
            "title": "Intervalseconds"
          },
          "nextRunAt": {
            "type": "string",
            "title": "Nextrunat"
          },
          "lastRunAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastrunat"
          },
          "lastStatus": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Laststatus"
          },
          "lastDetail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastdetail"
          },
          "lastScanId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastscanid"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "targetId",
          "scanProfile",
          "enabled",
          "intervalSeconds",
          "nextRunAt",
          "createdAt"
        ],
        "title": "ScheduleResponse",
        "description": "Schedule representation with run bookkeeping."
      },
      "ScheduleRunResponse": {
        "properties": {
          "scheduleId": {
            "type": "string",
            "format": "uuid",
            "title": "Scheduleid"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "default": ""
          },
          "scanId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scanid"
          }
        },
        "type": "object",
        "required": [
          "scheduleId",
          "status"
        ],
        "title": "ScheduleRunResponse",
        "description": "One executed tick: outcome only, safe to log."
      },
...
      "UpdateScheduleRequest": {
        "properties": {
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "intervalSeconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intervalseconds"
          },
          "scanProfile": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Scanprofile"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "UpdateScheduleRequest",
        "description": "PATCH /schedules/{id} body \u2014 any subset of mutable fields."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}